delete first version of implementation and replace with second
